Avoid empty non-nil slices and maps in repeated message fields
The JSON<->protobuf specification states:
  For repeated fields and map fields, they will be serialized if they are not empty.

In other words, an empty repeated field should produce no output.
However, the jsonpb marshaler currently marshals repeated fields as [] or {}
if the slice or map is non-nil. This is incorrect and needs to be fixed since
presence is not a concept for repeated fields in proto.

Modify this code to avoid setting non-nil lists and maps so that we can
land the proper jsonpb fix.

PiperOrigin-RevId: 266441145
Change-Id: I85a9d854582876111bd14546e9a5cc86c91e9ae3